isp_ex4_midterm/requirements.txt
appdirs==1.4.4
audioread==3.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
deepspeech==0.9.3
fonttools==4.40.0
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-speech==2.20.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jiwer==3.0.1
joblib==1.2.0
kiwisolver==1.4.4
language-tool-python==2.7.1
lazy-loader==0.2
Levenshtein==0.21.1
librosa==0.10.0.post2
llvmlite==0.40.1rc1
matplotlib==3.7.1
msgpack==1.0.5
nltk==3.8.1
noisereduce==2.0.1
numba==0.57.0
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pocketsphinx==5.0.1
pooch==1.6.0
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-Levenshtein==0.21.1
rapidfuzz==2.13.7
regex==2023.6.3
requests==2.31.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.5
spectral==0.23.1
SpeechRecognition==3.10.0
tabulate==0.9.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing-extensions==4.6.3
urllib3==1.26.16
webrtcvad==2.0.10
zipp==3.15.0
